test: split heap prof tests
The original heap prof tests can take too long to complete on
azure Windows machines, resulting in timeouts. Split them into
smaller tests and move them into the parallel directory to
speed up the execution.

PR-URL: https://github.com/nodejs/node/pull/44388
Refs: https://github.com/nodejs/reliability/issues/356
